Update sampleQC_UKB_MOD.r to reflect the update of UKB data
First 40 fields corresponding to first 40 PC's were added in the update
of UKB data, so the vector renaming the fields had to be updated accordingly.
In the future, the name of the fields should map to UKB data fields
rather than the name of the fields set to match the order of the UKB data fields.

This change does not affect the downstream data processing since the genetic_sex field
is still the thrid field in the updated sampleQC_UKB.tab. Also, sampleQC_UKB.tab was only
used in filering out sex mismatches in "04-exclude_sex_mismatch_subjects.Rmd".