Update the reformatting script for sampleQC_UKB.tab to reflect the update of UKB data
First 40 fields corresponding to first 40 PC's were added in the update
of UKB data, so the vector renaming the fields had to be updated accordingly.
In the future, the name of the fields should map to  UKB data fields
rather than naming of the fields based on the order of the UKB data fields.